Fixes #12: print friendlier message when switch partially succeeds (#13)
* Fixes #12: print friendlier message when switch partially succeeds

This change checks the exit code returned by the subprocess that
called `switch-to-configuration` on the target remote machine for a
partial success (i.e. `ExitFailure 4`). If we do receive a partial
success, then give a friendly message to the user indicating partial
success.

A partial success occurs when the configuration is switched and the
GRUB boot loader is updated but some of the systemd services fail to
start or restart.

* Add --failOnPartialSuccess as suggested by @intractable