Enable ssh keep live in issuing the reboot command. (#22)
When '--reboot' is given, `nix-deploy` will issue the reboot command via
ssh as the last step.  During reboot, systemd may terminate network interface
before the sshd service.  And there is no way for the `nix-deploy` to quickly
detect the connection loss.

To allow fast detection, this commit enables the ssh keep-live feature and
configure the interval to one second.